+++ packages/ui/src/state/index.ts
@@ -2,4 +2,5 @@ export * from "./globalUIStore";
 export * from "./notify";
 export * from "./themeStore";
 export * from "./themeColorScope";
+export * from "./ssrTheme";
 export * from "./SsrViewportContext";